Revert "Alpha18 vuex aggregation"
