+++ package.json
@@ -5,37 +5,36 @@
   "scripts": {
     "serve": "vue-cli-service serve",
     "dev": "vue-cli-service build --watch --mode production",
-    "watch": "vue-cli-service build --watch",
     "build": "vue-cli-service build",
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.30",
-    "@fortawesome/free-regular-svg-icons": "^5.14.0",
-    "@fortawesome/free-solid-svg-icons": "^5.14.0",
-    "@fortawesome/vue-fontawesome": "^0.1.10",
-    "axios": "^0.21.1",
-    "bootstrap": "^4.5.2",
-    "bootstrap-vue": "^2.16.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.28",
+    "@fortawesome/free-regular-svg-icons": "^5.13.0",
+    "@fortawesome/free-solid-svg-icons": "^5.13.0",
+    "@fortawesome/vue-fontawesome": "^0.1.9",
+    "axios": "^0.19.2",
+    "bootstrap": "^4.5.0",
+    "bootstrap-vue": "^2.15.0",
     "core-js": "^3.6.4",
-    "lodash": "^4.17.20",
+    "lodash": "^4.17.15",
     "textarea-caret": "^3.1.0",
-    "vue": "^2.6.12",
+    "vue": "^2.6.11",
     "vue-at": "^2.3.2",
     "vue-axios": "^2.1.5",
     "vue-configuration": "^1.0.3",
-    "vue-ctk-date-time-picker": "^2.5.0",
-    "vue-i18n": "^8.21.1",
+    "vue-ctk-date-time-picker": "^2.4.0",
+    "vue-i18n": "^8.18.2",
+    "vue-infinite-scroll": "^2.0.2",
     "vue-keep-scroll-position": "^0.1.2",
     "vue-lodash": "^2.1.2",
     "vue-moment": "^4.1.0",
     "vue-multiselect": "^2.1.6",
     "vue-nl2br": "^0.1.2",
-    "vue-spinner": "^1.0.4",
+    "vue-spinner": "^1.0.3",
     "vue-typeahead-bootstrap": "^2.6.0",
     "vue2-dropzone": "^3.6.0",
-    "vuedraggable": "^2.24.1",
-    "vuex": "^3.5.1"
+    "vuedraggable": "^2.23.2"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.3.0",
@@ -46,8 +45,7 @@
     "eslint-plugin-vue": "^6.2.2",
     "node-sass": "^4.14.1",
     "sass-loader": "^8.0.2",
-    "vue-template-compiler": "^2.6.12",
-    "vue-waypoint": "^3.5.2"
+    "vue-template-compiler": "^2.6.11"
   },
   "eslintConfig": {
     "root": true,
