+++ public/index.html
@@ -12,8 +12,7 @@
     <strong>We're sorry but <%= htmlWebpackPlugin.options.title %> doesn't work properly without JavaScript enabled.
         Please enable it to continue.</strong>
 </noscript>
-<div id="stream" data-container-nid="24330"></div>
-<!--<div id="stream"></div>-->
+<div id=stream data-container-nid=20821></div>
 <!-- built files will be auto injected -->
 </body>
 </html>
