+++ src/components/plugins/PluginMention/services/mention.service.js
@@ -3,7 +3,7 @@ import appConfig from '../../../../assets/config.js'
 
 export default {
 
-    async getMentionMembers(nid, cb, errorCb) {
+    async getMentionMembers(nid) {
         let apiNodeMentionUrl = appConfig.api.apiUserMentionMembers.replace('%node', nid)
 
         let requestData = {}
@@ -12,10 +12,10 @@ export default {
         let response = await Vue.axios.post(apiNodeMentionUrl, requestData, {withCredentials: true})
 
         if (response.data.status === 1) {
-            cb(response.data.mention_members)
+            return response.data.mention_members
         } else {
             // an error occured
-            errorCb()
+            alert(this.$t('warning.error_occured_please_repeat_your_action'))
         }
     }
 }
