+++ src/components/plugins/PluginPoll/services/poll.service.js
@@ -1,8 +1,8 @@
 import Vue from 'vue'
-import config from "@/assets/config";
+import appConfig from '../../../../assets/config.js'
 
 export default {
-    async addVote(nid, answerObj, token, cb, errorCb) {
+    async addVote(nid, answerObj, token) {
         let votes = {arSelected: []}
 
         if (answerObj.arSelected) {
@@ -11,15 +11,18 @@ export default {
             votes.arSelected = [{value: answerObj.value, votes: answerObj.votes}]
         }
 
-        let apiNodePollAddVoteUrl = config.api.apiNodePollAddVote.replace('%node', nid).replace('%token', token)
+        let apiNodePollAddVoteUrl = appConfig.api.apiNodePollAddVote.replace('%node', nid).replace('%token', token)
 
-        Vue.axios.post(apiNodePollAddVoteUrl, votes, {withCredentials: true}).then((response) => {
-            if (response.data.status === 1) {
-                return cb(response.data)
-            } else {
-                // an error occured
-                errorCb()
-            }
-        })
+        let requestData = votes.arSelected
+
+        // edit post => update
+        let response = await Vue.axios.post(apiNodePollAddVoteUrl, requestData, {withCredentials: true})
+
+        if (response.data.status === 1) {
+            return response.data
+        } else {
+            // an error occured
+            alert(this.$t('warning.error_occured_please_repeat_your_action'))
+        }
     }
 }
