--- src/helper/helper.js
@@ -1,5 +1,0 @@
-export default {
-    test: function () {
-        return "test"
-    },
-}
