+++ src/i18n/de_DE.json
@@ -18,9 +18,7 @@
       "customdictFileTooBig": "Diese Datei ist zu groß. Maximale Größe: {{maxFilesize}} MB.",
       "nmDictHideFile": "Verstecken",
       "dictMaxFilesExceeded": "Das Uploadlimit von {{maxFiles}} wurde überschritten",
-      "postContextFrom": "Beitrag aus",
-      "edited": "bearbeitet",
-      "load_more": "mehr laden..."
+      "postContextFrom": "Beitrag aus"
     },
     "link": {
       "show_x_comments": "{count} Kommentare anzeigen",
@@ -66,9 +64,6 @@
         "status_error": "Fehler",
         "vote": "Stimme",
         "votes": "Stimmen",
-        "participant": "Teilnahme",
-        "participants": "Teilnahmen",
-        "participated": "teilgenommen",
         "pollEndDateLabel": "Ende der Umfrage festlegen",
         "pollEndDateDescription": "Wird dieses Feld gesetzt, so ist ab diesem Zeitpunkt keine Abstimmung mehr möglich. Es wird das Umfrageergebnis angezeigt."
       }
