+++ vue.config.js
@@ -1,8 +1,5 @@
 module.exports = {
     filenameHashing: false,
-    //publicPath: '../onlinekurslabor/web/sites/all/modules/onlinekurslabor/nm_stream/libs/nm_stream/dist/',
-    //outputDir: '../onlinekurslabor/web/sites/all/modules/onlinekurslabor/nm_stream/libs/nm_stream/dist/',
-    configureWebpack: {
-        devtool: 'source-map'
-    }
+    publicPath: '../onlinekurslabor/web/sites/all/modules/onlinekurslabor/nm_stream/libs/nm_stream/dist/',
+    outputDir: '../onlinekurslabor/web/sites/all/modules/onlinekurslabor/nm_stream/libs/nm_stream/dist/',
 }
